pyproject.toml
[project]
name = "feishunanobot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.45.0,<1.0.0",
    "chardet>=3.0.2,<6.0.0",
    "croniter>=6.0.0,<7.0.0",
    "ddgs>=9.5.5,<10.0.0",
    "dingtalk-stream>=0.24.0,<1.0.0",
    "dulwich>=0.22.0,<1.0.0",
    "faiss-cpu>=1.13.2",
    "filelock>=3.25.2",
    "httpx>=0.28.0,<1.0.0",
    "jinja2>=3.1.0,<4.0.0",
    "json-repair>=0.57.0,<1.0.0",
    "jsonlines>=4.0.0",
    "lark-oapi>=1.5.0,<2.0.0",
    "lightrag-hku",
    "loguru>=0.7.3,<1.0.0",
    "mcp>=1.19.0",
    "modelscope>=1.36.0",
    "msgpack>=1.1.0,<2.0.0",
    "neo4j>=6.1.0",
    "numpy>=2.4.4",
    "oauth-cli-kit>=0.1.3,<1.0.0",
    "openai>=2.8.0",
    "openpyxl>=3.1.0,<4.0.0",
    "prompt-toolkit>=3.0.50,<4.0.0",
    "pydantic>=2.12.0,<3.0.0",
    "pydantic-settings>=2.12.0,<3.0.0",
    "pypdf>=6.10.2",
    "python-docx>=1.1.0,<2.0.0",
    "python-pptx>=1.0.0,<2.0.0",
    "python-socketio>=5.16.0,<6.0.0",
    "python-socks[asyncio]>=2.8.0,<3.0.0",
    "python-telegram-bot[socks]>=22.6,<23.0",
    "qq-botpy>=1.2.0,<2.0.0",
    "questionary>=2.0.0,<3.0.0",
    "readability-lxml>=0.8.4,<1.0.0",
    "rich>=14.0.0,<15.0.0",
    "sentence-transformers>=5.4.1",
    "slack-sdk>=3.39.0,<4.0.0",
    "slackify-markdown>=0.2.0,<1.0.0",
    "socksio>=1.0.0,<2.0.0",
    "tiktoken>=0.12.0,<1.0.0",
    "typer>=0.20.0,<1.0.0",
    "websocket-client>=1.9.0,<2.0.0",
    "websockets>=16.0,<17.0",
]

[tool.uv.workspace]
members = [
    "thirdparty/LightRAG",
]

[tool.uv.sources]
lightrag-hku = { workspace = true }

[dependency-groups]
dev = [
    "aiohttp>=3.9.0,<4.0.0",
    "pymupdf>=1.25.0",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0,<2.0.0",
    "pytest-cov>=6.0.0,<7.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
fsbot = "nanobot.cli.commands:app"

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["nanobot"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
